feat: Add functionality to add new todos
This commit adds functionality to the AddTodo component to allow users to add new todos. The handleSubmit function now captures the form data and adds a new todo to the list using the addTodo function from the RootContextProvider. The form is then reset after the todo is added.

Refactor the code to remove unnecessary console.log statements and improve code readability.